requirements.txt
python-telegram-bot==20.7
youtube_dl